requirements.txt
attrs==23.1.0
build==0.10.0
CacheControl==0.12.14
certifi==2023.5.7
charset-normalizer==3.1.0
cleo==2.0.1
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.6
dulwich==0.21.5
environs==9.5.0
filelock==3.12.2
grpcio==1.53.0
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
installer==0.7.0
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
marshmallow==3.19.0
more-itertools==9.1.0
msgpack==1.0.5
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pexpect==4.8.0
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.5.3
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
protobuf==4.23.3
ptyprocess==0.7.0
pymilvus==2.2.12
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32-ctypes==0.2.0
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.0.post1
six==1.16.0
tomli==2.0.1
tomlkit==0.11.8
trove-classifiers==2023.5.24
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.16
virtualenv==20.23.0
webencodings==0.5.1
wxPython==4.2.1
zipp==3.15.0
